wallet_project/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
psycopg2-binary==2.9.3
